MSA-10578 - Detect when upgrade fails.
install.sh checks the version of MSA after the images are upgraded. If
anything else fails, trying to rerun the script will cause the
check to falsely think that there is no upgrade to do.

A file (.upgrade_unfinished) will be created in the beggining of the
installation and removed after installation is sucessful. If
installation fails in the middle, reruning the script will check for
that file, that, if present, will represent an unfinished installation.